Parse API request body only for API requests
The API middleware parsed the request body for all incoming requests to
a Javascript Object and persisted it in `req.body`.
This broke downstream proxy functionality that relies on the body being
a string.

This fix ensures that the API request handler only persists the body as
an object when it's actually called, by testing the route.

Resolves #59.