Remove dodgy request parsing
In record mode the request payload was being parsed by a call to
`decodeURIComponent` which assumes the payload is URI-encoded.

This is not always true, and may throw exceptions if the payload
contains illegal characters. Surprisingly this is apparently not
that common, but definitely happens with binary payloads.

In addition, this logic was not apparently used anywhere in the
connect-prism code itself, and was added as a convenience by a
contributor. This functionality is best provided via middleware
anyway